Reapply "cheri: Return bounded capabilities from the page allocator"
This reverts commit a8a9f126970ae0df491c31ab5cf8d05431cd18e5.

While there add some minor improvements:
- be more carefule with argument handling in the page_to_virt() macro.
- Initialize page alloc order in mm_init.c for reserved pages.

+++ mm/kasan/shadow.c
@@ -397,9 +397,9 @@ int kasan_populate_vmalloc(unsigned long addr, unsigned long size)
 static int kasan_depopulate_vmalloc_pte(pte_t *ptep, unsigned long addr,
 					void *unused)
 {
-	unsigned long page;
+	uintptr_t page;
 
-	page = (unsigned long)__va(pte_pfn(ptep_get(ptep)) << PAGE_SHIFT);
+	page = (uintptr_t)__va(pte_pfn(ptep_get(ptep)) << PAGE_SHIFT);
 
 	spin_lock(&init_mm.page_table_lock);
